requirements.txt
accelerate==1.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
diffusers==0.35.2
filelock==3.19.1
fsspec==2025.9.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
psutil==7.1.3
pygame==2.6.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
setuptools==70.2.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
